+++ python/fast_mlsirm/_legacy_init.py
@@ -64,6 +64,7 @@ from .nominal import fit_nominal as fit_nominal, NominalResponseFit as NominalRe
 from .grm import fit_grm as fit_grm, GrmFit as GrmFit
 from .bifactor_grm import fit_bifactor_grm as fit_bifactor_grm, BifactorGrmFit as BifactorGrmFit
 from .bifactor_multigroup import fit_bifactor_grm_multigroup as fit_bifactor_grm_multigroup, BifactorMultigroupFit as BifactorMultigroupFit
+from .two_tier_grm import fit_two_tier_grm as fit_two_tier_grm, TwoTierGrmFit as TwoTierGrmFit
 from .gpcm import fit_gpcm as fit_gpcm, GpcmFit as GpcmFit
 from .facets import fit_facets as fit_facets, FacetsFit as FacetsFit
 from .ksirt import ksirt_analysis as ksirt_analysis, KsirtResult as KsirtResult
@@ -394,6 +395,8 @@ __all__ = [  # noqa: RUF022
     "BifactorGrmFit",
     "fit_bifactor_grm_multigroup",
     "BifactorMultigroupFit",
+    "fit_two_tier_grm",
+    "TwoTierGrmFit",
     "fit_gpcm",
     "GpcmFit",
     "fit_rsm",